+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## UNRELEASED
+
+* Issue 252: Check for legitimate values in JSON date-time fields by including
+  a `FormatChecker` instance when running `jsonschema` validation.
+
 ## v1.18.0 (2026-09-22)
 
 * Minor version release for Issue 323 changes.